use GLubyte for stencil ref/mask/clear

+++ src/mesa/softpipe/sp_state.h
@@ -131,10 +131,10 @@ struct softpipe_stencil_state {
    GLuint back_fail_op:3;
    GLuint back_zpass_op:3;
    GLuint back_zfail_op:3;
-   GLint ref_value[2];      /**< [0] = front, [1] = back */
-   GLuint value_mask[2];
-   GLuint write_mask[2];
-   GLuint clear_value;
+   GLubyte ref_value[2];      /**< [0] = front, [1] = back */
+   GLubyte value_mask[2];
+   GLubyte write_mask[2];
+   GLubyte clear_value;
 };
